requirements.txt
scikit-learn==1.3.0
pandas==1.1.3
numpy==1.24.4
joblib==1.3.1
mysql-connector-python==8.0.31
pywhatkit==5.4
PyAutoGUI==0.9.54
selenium==4.10.0
SQLAlchemy==1.4.43
schedule==1.2.0